+++ sdk/iothub/azure-mgmt-iothub/azure/mgmt/iothub/_iot_hub_client.py
@@ -46,7 +46,7 @@ class IotHubClient(MultiApiClientMixin, SDKClient):
     :type profile: azure.profiles.KnownProfiles
     """
 
-    DEFAULT_API_VERSION = '2019-11-04'
+    DEFAULT_API_VERSION = '2020-03-01'
     _PROFILE_TAG = "azure.mgmt.iothub.IotHubClient"
     LATEST_PROFILE = ProfileDefinition({
         _PROFILE_TAG: {
@@ -82,6 +82,7 @@ class IotHubClient(MultiApiClientMixin, SDKClient):
            * 2019-03-22-preview: :mod:`v2019_03_22_preview.models<azure.mgmt.iothub.v2019_03_22_preview.models>`
            * 2019-07-01-preview: :mod:`v2019_07_01_preview.models<azure.mgmt.iothub.v2019_07_01_preview.models>`
            * 2019-11-04: :mod:`v2019_11_04.models<azure.mgmt.iothub.v2019_11_04.models>`
+           * 2020-03-01: :mod:`v2020_03_01.models<azure.mgmt.iothub.v2020_03_01.models>`
         """
         if api_version == '2016-02-03':
             from .v2016_02_03 import models
@@ -113,6 +114,9 @@ class IotHubClient(MultiApiClientMixin, SDKClient):
         elif api_version == '2019-11-04':
             from .v2019_11_04 import models
             return models
+        elif api_version == '2020-03-01':
+            from .v2020_03_01 import models
+            return models
         raise NotImplementedError("APIVersion {} is not available".format(api_version))
 
     @property
@@ -127,6 +131,7 @@ class IotHubClient(MultiApiClientMixin, SDKClient):
            * 2019-03-22-preview: :class:`CertificatesOperations<azure.mgmt.iothub.v2019_03_22_preview.operations.CertificatesOperations>`
            * 2019-07-01-preview: :class:`CertificatesOperations<azure.mgmt.iothub.v2019_07_01_preview.operations.CertificatesOperations>`
            * 2019-11-04: :class:`CertificatesOperations<azure.mgmt.iothub.v2019_11_04.operations.CertificatesOperations>`
+           * 2020-03-01: :class:`CertificatesOperations<azure.mgmt.iothub.v2020_03_01.operations.CertificatesOperations>`
         """
         api_version = self._get_api_version('certificates')
         if api_version == '2017-07-01':
@@ -145,6 +150,8 @@ class IotHubClient(MultiApiClientMixin, SDKClient):
             from .v2019_07_01_preview.operations import CertificatesOperations as OperationClass
         elif api_version == '2019-11-04':
             from .v2019_11_04.operations import CertificatesOperations as OperationClass
+        elif api_version == '2020-03-01':
+            from .v2020_03_01.operations import CertificatesOperations as OperationClass
         else:
             raise NotImplementedError("APIVersion {} is not available".format(api_version))
         return OperationClass(self._client, self.config, Serializer(self._models_dict(api_version)), Deserializer(self._models_dict(api_version)))
@@ -157,6 +164,7 @@ class IotHubClient(MultiApiClientMixin, SDKClient):
            * 2019-03-22-preview: :class:`IotHubOperations<azure.mgmt.iothub.v2019_03_22_preview.operations.IotHubOperations>`
            * 2019-07-01-preview: :class:`IotHubOperations<azure.mgmt.iothub.v2019_07_01_preview.operations.IotHubOperations>`
            * 2019-11-04: :class:`IotHubOperations<azure.mgmt.iothub.v2019_11_04.operations.IotHubOperations>`
+           * 2020-03-01: :class:`IotHubOperations<azure.mgmt.iothub.v2020_03_01.operations.IotHubOperations>`
         """
         api_version = self._get_api_version('iot_hub')
         if api_version == '2019-03-22':
@@ -167,6 +175,8 @@ class IotHubClient(MultiApiClientMixin, SDKClient):
             from .v2019_07_01_preview.operations import IotHubOperations as OperationClass
         elif api_version == '2019-11-04':
             from .v2019_11_04.operations import IotHubOperations as OperationClass
+        elif api_version == '2020-03-01':
+            from .v2020_03_01.operations import IotHubOperations as OperationClass
         else:
             raise NotImplementedError("APIVersion {} is not available".format(api_version))
         return OperationClass(self._client, self.config, Serializer(self._models_dict(api_version)), Deserializer(self._models_dict(api_version)))
@@ -185,6 +195,7 @@ class IotHubClient(MultiApiClientMixin, SDKClient):
            * 2019-03-22-preview: :class:`IotHubResourceOperations<azure.mgmt.iothub.v2019_03_22_preview.operations.IotHubResourceOperations>`
            * 2019-07-01-preview: :class:`IotHubResourceOperations<azure.mgmt.iothub.v2019_07_01_preview.operations.IotHubResourceOperations>`
            * 2019-11-04: :class:`IotHubResourceOperations<azure.mgmt.iothub.v2019_11_04.operations.IotHubResourceOperations>`
+           * 2020-03-01: :class:`IotHubResourceOperations<azure.mgmt.iothub.v2020_03_01.operations.IotHubResourceOperations>`
         """
         api_version = self._get_api_version('iot_hub_resource')
         if api_version == '2016-02-03':
@@ -207,6 +218,8 @@ class IotHubClient(MultiApiClientMixin, SDKClient):
             from .v2019_07_01_preview.operations import IotHubResourceOperations as OperationClass
         elif api_version == '2019-11-04':
             from .v2019_11_04.operations import IotHubResourceOperations as OperationClass
+        elif api_version == '2020-03-01':
+            from .v2020_03_01.operations import IotHubResourceOperations as OperationClass
         else:
             raise NotImplementedError("APIVersion {} is not available".format(api_version))
         return OperationClass(self._client, self.config, Serializer(self._models_dict(api_version)), Deserializer(self._models_dict(api_version)))
@@ -223,6 +236,7 @@ class IotHubClient(MultiApiClientMixin, SDKClient):
            * 2019-03-22-preview: :class:`Operations<azure.mgmt.iothub.v2019_03_22_preview.operations.Operations>`
            * 2019-07-01-preview: :class:`Operations<azure.mgmt.iothub.v2019_07_01_preview.operations.Operations>`
            * 2019-11-04: :class:`Operations<azure.mgmt.iothub.v2019_11_04.operations.Operations>`
+           * 2020-03-01: :class:`Operations<azure.mgmt.iothub.v2020_03_01.operations.Operations>`
         """
         api_version = self._get_api_version('operations')
         if api_version == '2017-07-01':
@@ -241,6 +255,34 @@ class IotHubClient(MultiApiClientMixin, SDKClient):
             from .v2019_07_01_preview.operations import Operations as OperationClass
         elif api_version == '2019-11-04':
             from .v2019_11_04.operations import Operations as OperationClass
+        elif api_version == '2020-03-01':
+            from .v2020_03_01.operations import Operations as OperationClass
+        else:
+            raise NotImplementedError("APIVersion {} is not available".format(api_version))
+        return OperationClass(self._client, self.config, Serializer(self._models_dict(api_version)), Deserializer(self._models_dict(api_version)))
+
+    @property
+    def private_endpoint_connections(self):
+        """Instance depends on the API version:
+
+           * 2020-03-01: :class:`PrivateEndpointConnectionsOperations<azure.mgmt.iothub.v2020_03_01.operations.PrivateEndpointConnectionsOperations>`
+        """
+        api_version = self._get_api_version('private_endpoint_connections')
+        if api_version == '2020-03-01':
+            from .v2020_03_01.operations import PrivateEndpointConnectionsOperations as OperationClass
+        else:
+            raise NotImplementedError("APIVersion {} is not available".format(api_version))
+        return OperationClass(self._client, self.config, Serializer(self._models_dict(api_version)), Deserializer(self._models_dict(api_version)))
+
+    @property
+    def private_link_resources(self):
+        """Instance depends on the API version:
+
+           * 2020-03-01: :class:`PrivateLinkResourcesOperations<azure.mgmt.iothub.v2020_03_01.operations.PrivateLinkResourcesOperations>`
+        """
+        api_version = self._get_api_version('private_link_resources')
+        if api_version == '2020-03-01':
+            from .v2020_03_01.operations import PrivateLinkResourcesOperations as OperationClass
         else:
             raise NotImplementedError("APIVersion {} is not available".format(api_version))
         return OperationClass(self._client, self.config, Serializer(self._models_dict(api_version)), Deserializer(self._models_dict(api_version)))
@@ -255,6 +297,7 @@ class IotHubClient(MultiApiClientMixin, SDKClient):
            * 2019-03-22-preview: :class:`ResourceProviderCommonOperations<azure.mgmt.iothub.v2019_03_22_preview.operations.ResourceProviderCommonOperations>`
            * 2019-07-01-preview: :class:`ResourceProviderCommonOperations<azure.mgmt.iothub.v2019_07_01_preview.operations.ResourceProviderCommonOperations>`
            * 2019-11-04: :class:`ResourceProviderCommonOperations<azure.mgmt.iothub.v2019_11_04.operations.ResourceProviderCommonOperations>`
+           * 2020-03-01: :class:`ResourceProviderCommonOperations<azure.mgmt.iothub.v2020_03_01.operations.ResourceProviderCommonOperations>`
         """
         api_version = self._get_api_version('resource_provider_common')
         if api_version == '2018-04-01':
@@ -269,6 +312,8 @@ class IotHubClient(MultiApiClientMixin, SDKClient):
             from .v2019_07_01_preview.operations import ResourceProviderCommonOperations as OperationClass
         elif api_version == '2019-11-04':
             from .v2019_11_04.operations import ResourceProviderCommonOperations as OperationClass
+        elif api_version == '2020-03-01':
+            from .v2020_03_01.operations import ResourceProviderCommonOperations as OperationClass
         else:
             raise NotImplementedError("APIVersion {} is not available".format(api_version))
         return OperationClass(self._client, self.config, Serializer(self._models_dict(api_version)), Deserializer(self._models_dict(api_version)))